Fixing UpdatePolicy validation
